refactor(runner): extract session history restore planning
