+++ crates/runner/src/cmd/start/job_discovery.rs
@@ -24,11 +24,9 @@ use super::idle_lifecycle::{
 use super::job_spawn::{JobProfile, SpawnContext, SpawnJobRequest, spawn_job};
 use crate::config::ProfileConfig;
 use crate::executor::{
-    RunnerPreSpawnPhase, RunnerPreSpawnTiming, SessionHistoryCpuPool, SessionHistoryMaterializer,
-    SessionHistoryProbe, SessionHistoryRestoreFallback, SessionHistoryRestorePlan,
-    effective_cli_framework, validate_resume_session_id,
+    RunnerPreSpawnPhase, RunnerPreSpawnTiming, SessionHistoryRestorePlanInput,
+    build_session_history_restore_plan, validate_resume_session_id,
 };
-use crate::http::HttpClient;
 use crate::idle_pool::{
     DestroyOutcome, IdlePoolSnapshot, IdleUnparkResult, ReservedIdleSandbox,
     RestoreReservedIdleResult, ReusableIdleSandbox,
@@ -39,9 +37,6 @@ use crate::provider::{
     ClaimedJob, JobCandidate, LocalAdmissionResourceKind, SessionAffinityLocalResource,
 };
 use crate::resource_budget::{BudgetLease, ResourceBudget};
-use crate::restored_session_identity::{
-    RestoredSessionIdentity, RestoredSessionIdentityMismatchReason,
-};
 use crate::run_cancellation::{RunCancellationHandle, SharedRunCancellationMap};
 use crate::status::{RunnerMode, StatusTracker};
 use crate::types::{
@@ -260,22 +255,20 @@ pub(super) async fn handle_discovered_job(
             }
         };
 
-    let session_history_restore_plan = if resume_session_valid {
-        build_session_history_restore_plan(
-            &ctx.spawn_ctx.exec_config.http,
-            &ctx.spawn_ctx.exec_config.session_history_cpu,
-            claimed.context(),
-            &job_cancel,
-            SessionHistoryRestoreReuse {
-                entry: reuse_entry.as_ref(),
-                result: reuse_result,
-            },
-            &mut pre_spawn_timing,
-            Some(&ctx.spawn_ctx.exec_config.session_history_probe),
-        )
-    } else {
-        SessionHistoryRestorePlan::Default
-    };
+    let session_history_restore_plan =
+        build_session_history_restore_plan(SessionHistoryRestorePlanInput {
+            resume_session_valid,
+            http: &ctx.spawn_ctx.exec_config.http,
+            cpu: &ctx.spawn_ctx.exec_config.session_history_cpu,
+            context: claimed.context(),
+            cancel: job_cancel.token(),
+            reuse_result,
+            restored_identity: reuse_entry
+                .as_ref()
+                .and_then(ReusableIdleSandbox::restored_session_identity),
+            pre_spawn_timing: &mut pre_spawn_timing,
+            probe: Some(&ctx.spawn_ctx.exec_config.session_history_probe),
+        });
 
     // Determine sandbox_id after the reuse decision. On reuse, the sandbox keeps
     // its original identity; on a fresh create, allocate a new UUID for the
@@ -321,109 +314,6 @@ pub(super) async fn handle_discovered_job(
         ctx.jobs,
     );
     needs_session_affinity_refresh
-}
-
-struct SessionHistoryRestoreReuse<'a> {
-    entry: Option<&'a ReusableIdleSandbox>,
-    result: SandboxReuseResult,
-}
-
-fn build_session_history_restore_plan(
-    http: &HttpClient,
-    cpu: &SessionHistoryCpuPool,
-    context: &ExecutionContext,
-    cancel: &RunCancellationHandle,
-    reuse: SessionHistoryRestoreReuse<'_>,
-    pre_spawn_timing: &mut RunnerPreSpawnTiming,
-    probe: Option<&SessionHistoryProbe>,
-) -> SessionHistoryRestorePlan {
-    let Some(resume_session) = context.resume_session.as_ref() else {
-        return SessionHistoryRestorePlan::Default;
-    };
-    if resume_session.history_ref().is_none() {
-        return SessionHistoryRestorePlan::Default;
-    }
-
-    let mut prefix_attribution = None;
-    let fallback = match reuse.result {
-        SandboxReuseResult::Reused => {
-            let requested_identity = RestoredSessionIdentity::from_context(context);
-            if let Some(requested_identity) = requested_identity {
-                match reuse
-                    .entry
-                    .and_then(ReusableIdleSandbox::restored_session_identity)
-                {
-                    Some(restored_identity)
-                        if restored_identity.is_verified_match_for_request(&requested_identity) =>
-                    {
-                        return SessionHistoryRestorePlan::SkipVerified(restored_identity.clone());
-                    }
-                    Some(restored_identity) if restored_identity == &requested_identity => {
-                        if restored_identity.has_final_metadata_verification() {
-                            Some(SessionHistoryRestoreFallback::IdentityMismatch(
-                                restored_identity.mismatch_reason_for_request(&requested_identity),
-                            ))
-                        } else {
-                            Some(SessionHistoryRestoreFallback::UnverifiedIdleIdentity)
-                        }
-                    }
-                    Some(restored_identity) => {
-                        let (mismatch_reason, attribution) = restored_identity
-                            .mismatch_reason_and_prefix_attribution(&requested_identity);
-                        prefix_attribution = attribution;
-                        Some(SessionHistoryRestoreFallback::IdentityMismatch(
-                            mismatch_reason,
-                        ))
-                    }
-                    None => Some(SessionHistoryRestoreFallback::MissingIdleIdentity),
-                }
-            } else {
-                Some(SessionHistoryRestoreFallback::IdentityMismatch(Some(
-                    RestoredSessionIdentityMismatchReason::MissingRequestedIdentity,
-                )))
-            }
-        }
-        SandboxReuseResult::NoSessionId
-        | SandboxReuseResult::PoolMiss
-        | SandboxReuseResult::ProfileMismatch
-        | SandboxReuseResult::DeviceLimitMismatch
-        | SandboxReuseResult::UnparkFailed => Some(SessionHistoryRestoreFallback::NonReuse),
-    };
-
-    if reuse.result != SandboxReuseResult::Reused {
-        return SessionHistoryRestorePlan::DeferredHashBacked { fallback };
-    }
-
-    let started_at = Instant::now();
-    let materializer = match prefix_attribution {
-        Some(prefix_attribution) => {
-            SessionHistoryMaterializer::start_cancellable_with_prefix_attribution(
-                http,
-                cpu,
-                Some(resume_session),
-                effective_cli_framework(&context.cli_agent_type),
-                cancel.token(),
-                probe,
-                prefix_attribution,
-            )
-        }
-        None => SessionHistoryMaterializer::start_cancellable(
-            http,
-            cpu,
-            Some(resume_session),
-            effective_cli_framework(&context.cli_agent_type),
-            cancel.token(),
-            probe,
-        ),
-    };
-    pre_spawn_timing.record_phase_elapsed(
-        RunnerPreSpawnPhase::SessionHistoryMaterializerStart,
-        started_at,
-    );
-    SessionHistoryRestorePlan::Prestarted {
-        materializer,
-        fallback,
-    }
 }
 
 async fn publish_active_run_status(
@@ -1188,31 +1078,7 @@ async fn try_reuse_from_pool(
 mod tests {
     use super::*;
 
-    use super::super::StartLoopTestObserver;
-    use super::super::job_lifecycle::{ActiveBudgetLease, CompletionPayload, RunCleanupState};
-    use super::super::sandbox_finalization::{FinalizeContext, finalize_sandbox_for_completion};
-    use crate::http::HttpClientConfig;
-    use crate::idle_pool::test_support::ParkedIdleCandidateBuilder;
-    use crate::idle_pool::{IdlePool, IdlePoolConfig, ParkResult, ParkingGate};
-    use crate::idle_reuse_preparation::mock_sandbox_ready_for_idle_reuse;
-    use crate::network_log_drain::NetworkLogDrainCoordinator;
-    use crate::provider::CompletionAuth;
-    use crate::resource_budget::ResourceBudget;
-    use crate::restored_session_identity::{
-        RestoredSessionFramework, RestoredSessionHistoryHashSizeRelationship,
-    };
     use crate::status::IdleVm;
-    use crate::test_fixtures::execution_context_for_test;
-    use crate::types::{ResumeSession, ResumeSessionHistory, ResumeSessionHistoryRef};
-    use guest_contracts::{
-        codex_thread_id::canonical_codex_thread_id,
-        session_history_identity::{
-            FinalSessionHistoryFramework, FinalSessionHistoryIdentity, FinalSessionHistoryRefKind,
-        },
-    };
-    use sandbox::SandboxFactory;
-    use sandbox_mock::MockSandboxFactory;
-    use sha2::{Digest, Sha256};
 
     fn read_active_run_phase(path: &std::path::Path) -> String {
         let raw = std::fs::read_to_string(path).unwrap();
@@ -1231,160 +1097,6 @@ mod tests {
                 sandbox_id: SandboxId::new_v4(),
             }],
         }
-    }
-
-    fn test_http_client() -> HttpClient {
-        HttpClient::new(HttpClientConfig {
-            api_url: "http://localhost".into(),
-            vercel_bypass: None,
-            client_session_id: "runner-session-test".to_string(),
-        })
-        .unwrap()
-    }
-
-    fn context_with_history_ref(history_hash: &str) -> ExecutionContext {
-        context_with_history_ref_and_size(history_hash, 12)
-    }
-
-    fn context_with_history_ref_and_size(history_hash: &str, size: u64) -> ExecutionContext {
-        let mut context = execution_context_for_test(RunId::new_v4());
-        context.resume_session = Some(ResumeSession {
-            cli_agent_session_id: "sess-restore-plan".into(),
-            history: ResumeSessionHistory::Ref {
-                history_ref: ResumeSessionHistoryRef {
-                    kind: crate::types::ResumeSessionHistoryRefKind::Blob,
-                    hash: history_hash.into(),
-                    url: "http://127.0.0.1:9/history.blob".into(),
-                    encoding: None,
-                    raw_size: size,
-                    encoded_size: size,
-                    download_source: None,
-                },
-            },
-        });
-        context
-    }
-
-    fn final_metadata_identity(history_hash: String, size: u64) -> RestoredSessionIdentity {
-        let metadata = FinalSessionHistoryIdentity::new(
-            FinalSessionHistoryFramework::ClaudeCode,
-            hex::encode(Sha256::digest(b"sess-restore-plan")),
-            FinalSessionHistoryRefKind::Blob,
-            history_hash,
-            size,
-            "/home/user/.claude/projects/-home-user-workspace/session.jsonl",
-        )
-        .unwrap();
-        RestoredSessionIdentity::from_final_metadata(
-            metadata,
-            "/home/user/.vm0/guest-agent/runs/previous/final-session-history-identity.json",
-            "/home/user/.vm0/guest-agent/runs/previous",
-        )
-        .expect("checkpointed final identity")
-    }
-
-    async fn reusable_sandbox_with_identity(
-        restored_session_identity: Option<RestoredSessionIdentity>,
-    ) -> ReusableIdleSandbox {
-        let budget = Arc::new(ResourceBudget::new(1, 1, 1.0, 0));
-        let lease = ResourceBudget::try_reserve_lease(&budget, 1, 1).unwrap();
-        let builder = ParkedIdleCandidateBuilder::new("sess-restore-plan", lease);
-        let builder = if let Some(restored_session_identity) = restored_session_identity {
-            builder.with_restored_session_identity(restored_session_identity)
-        } else {
-            builder
-        };
-        let candidate = builder.build();
-        let mut pool = IdlePool::new(IdlePoolConfig {
-            default_timeout: std::time::Duration::from_secs(300),
-            max_idle: 0,
-        });
-        assert!(matches!(pool.park(candidate), ParkResult::Parked));
-        let entry = pool
-            .take("sess-restore-plan")
-            .expect("idle entry should exist");
-        let IdleUnparkResult::Reused { sandbox, .. } = entry.try_unpark().await else {
-            panic!("idle entry should unpark");
-        };
-        *sandbox
-    }
-
-    async fn reusable_sandbox_parked_by_finalizer(
-        restored_session_identity: Option<RestoredSessionIdentity>,
-    ) -> ReusableIdleSandbox {
-        let dir = tempfile::tempdir().unwrap();
-        let status = Arc::new(StatusTracker::new(
-            dir.path().join("status.json"),
-            4,
-            None,
-            None,
-        ));
-        status.write_initial().await;
-        let parking_gate = ParkingGate::new_open();
-        let idle_pool: SharedIdlePool =
-            Arc::new(tokio::sync::Mutex::new(IdlePool::new_with_parking_gate(
-                IdlePoolConfig {
-                    default_timeout: std::time::Duration::from_secs(300),
-                    max_idle: 10,
-                },
-                parking_gate.clone(),
-            )));
-        let budget = Arc::new(ResourceBudget::new(2, 4096, 1.0, 0));
-        let lease = ResourceBudget::try_reserve_lease(&budget, 2, 4096).unwrap();
-        let run_id = RunId::new_v4();
-        let sandbox_id = SandboxId::new_v4();
-
-        let _completion_ready = finalize_sandbox_for_completion(
-            Some(Box::new(mock_sandbox_ready_for_idle_reuse(
-                "restore-plan-finalizer",
-            ))),
-            ActiveBudgetLease::new(lease),
-            CompletionPayload::new(
-                run_id,
-                0,
-                None,
-                sandbox_id,
-                SandboxReuseResult::PoolMiss,
-                CompletionAuth::local(),
-            ),
-            FinalizeContext {
-                run_id,
-                sandbox_id,
-                profile_name: "vm0/default".into(),
-                cli_agent_session_id: Some("sess-restore-plan".into()),
-                discovered_cli_agent_session_id: None,
-                restored_session_identity,
-                source_ip: "10.0.0.1".into(),
-                network_log_session: None,
-                workspace_image: None,
-                workspace_image_size_bytes: 0,
-                storage_fingerprints: crate::storage_fingerprints::StorageFingerprints::default(),
-                device_rate_limits: None,
-                factory: Arc::new(Box::new(MockSandboxFactory::new()) as Box<dyn SandboxFactory>),
-                idle_pool: Arc::clone(&idle_pool),
-                status,
-                park_notify: Arc::new(tokio::sync::Notify::new()),
-                held_session_snapshot: super::super::heartbeat::HeldSessionStateSnapshot::new(),
-                parking_gate,
-                network_log_drain: NetworkLogDrainCoordinator::noop(),
-                exit_code: 0,
-                cancel: RunCancellationHandle::new(),
-                cleanup_state: RunCleanupState::new(),
-                outer_job_panic: None,
-                test_observer: StartLoopTestObserver::default(),
-            },
-        )
-        .await;
-
-        let entry = idle_pool
-            .lock()
-            .await
-            .take("sess-restore-plan")
-            .expect("finalizer should park reusable sandbox");
-        let IdleUnparkResult::Reused { sandbox, .. } = entry.try_unpark().await else {
-            panic!("finalized idle entry should unpark");
-        };
-        *sandbox
     }
 
     #[tokio::test]
@@ -1421,521 +1133,5 @@ mod tests {
         .await;
 
         assert_eq!(read_active_run_phase(&status_path), "running");
-    }
-
-    #[tokio::test]
-    async fn restore_plan_skips_matching_checkpointed_final_identity() {
-        let http = test_http_client();
-        let history_hash = "a".repeat(64);
-        let context = context_with_history_ref_and_size(&history_hash, 12);
-        let metadata_path =
-            "/home/user/.vm0/guest-agent/runs/previous/final-session-history-identity.json";
-        let runtime_dir = "/home/user/.vm0/guest-agent/runs/previous";
-        let metadata = FinalSessionHistoryIdentity::new(
-            FinalSessionHistoryFramework::ClaudeCode,
-            hex::encode(Sha256::digest(b"sess-restore-plan")),
-            FinalSessionHistoryRefKind::Blob,
-            history_hash,
-            12,
-            "/home/user/.claude/projects/-home-user-workspace/session.jsonl",
-        )
-        .unwrap();
-        let restored_identity =
-            RestoredSessionIdentity::from_final_metadata(metadata, metadata_path, runtime_dir)
-                .expect("checkpointed final identity");
-        let reusable_sandbox =
-            reusable_sandbox_with_identity(Some(restored_identity.clone())).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::SkipVerified(identity) => {
-                assert_eq!(identity, restored_identity);
-                assert_eq!(identity.history_size_bytes(), Some(12));
-                assert_eq!(identity.final_metadata_path(), Some(metadata_path));
-            }
-            _ => panic!("matching checkpointed final identity should skip restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_skips_matching_codex_checkpointed_final_identity() {
-        let http = test_http_client();
-        let history_hash = "a".repeat(64);
-        let mut context = execution_context_for_test(RunId::new_v4());
-        context.cli_agent_type = "codex".into();
-        context.resume_session = Some(ResumeSession {
-            cli_agent_session_id: "019E9154C30470F0ADDE36EFB1BE1701".into(),
-            history: ResumeSessionHistory::Ref {
-                history_ref: ResumeSessionHistoryRef {
-                    kind: crate::types::ResumeSessionHistoryRefKind::Blob,
-                    hash: history_hash.clone(),
-                    url: "http://127.0.0.1:9/history.blob".into(),
-                    encoding: None,
-                    raw_size: 12,
-                    encoded_size: 12,
-                    download_source: None,
-                },
-            },
-        });
-        let canonical_thread_id =
-            canonical_codex_thread_id("019E9154C30470F0ADDE36EFB1BE1701").unwrap();
-        let metadata_path =
-            "/home/user/.vm0/guest-agent/runs/previous/final-session-history-identity.json";
-        let runtime_dir = "/home/user/.vm0/guest-agent/runs/previous";
-        let metadata = FinalSessionHistoryIdentity::new(
-            FinalSessionHistoryFramework::Codex,
-            hex::encode(Sha256::digest(canonical_thread_id.as_bytes())),
-            FinalSessionHistoryRefKind::Blob,
-            history_hash,
-            12,
-            format!("CODEX_SEARCH:26:/home/user/.codex/sessions:{canonical_thread_id}"),
-        )
-        .unwrap();
-        let restored_identity =
-            RestoredSessionIdentity::from_final_metadata(metadata, metadata_path, runtime_dir)
-                .expect("checkpointed final identity");
-        let reusable_sandbox =
-            reusable_sandbox_with_identity(Some(restored_identity.clone())).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::SkipVerified(identity) => {
-                assert_eq!(identity, restored_identity);
-                assert_eq!(identity.history_size_bytes(), Some(12));
-                assert_eq!(identity.final_metadata_path(), Some(metadata_path));
-            }
-            _ => panic!("matching Codex checkpointed final identity should skip restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_skips_identity_parked_by_finalizer() {
-        let http = test_http_client();
-        let history_hash = "a".repeat(64);
-        let context = context_with_history_ref(&history_hash);
-        let metadata = FinalSessionHistoryIdentity::new(
-            FinalSessionHistoryFramework::ClaudeCode,
-            hex::encode(Sha256::digest(b"sess-restore-plan")),
-            FinalSessionHistoryRefKind::Blob,
-            history_hash,
-            12,
-            "/home/user/.claude/projects/-home-user-workspace/session.jsonl",
-        )
-        .unwrap();
-        let restored_identity = RestoredSessionIdentity::from_final_metadata(
-            metadata,
-            "/home/user/.vm0/guest-agent/runs/previous/final-session-history-identity.json",
-            "/home/user/.vm0/guest-agent/runs/previous",
-        )
-        .expect("checkpointed final identity");
-        let reusable_sandbox =
-            reusable_sandbox_parked_by_finalizer(Some(restored_identity.clone())).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::SkipVerified(identity) => {
-                assert_eq!(identity, restored_identity);
-            }
-            _ => panic!("finalizer-parked restored identity should skip restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_falls_back_when_matching_reused_identity_is_unverified() {
-        let http = test_http_client();
-        let context = context_with_history_ref("history-hash-a");
-        let restored_identity = RestoredSessionIdentity::from_context(&context).unwrap();
-        let reusable_sandbox = reusable_sandbox_with_identity(Some(restored_identity)).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::UnverifiedIdleIdentity)
-                );
-            }
-            _ => panic!("unverified reused identity should fall back to restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_falls_back_when_matching_reused_identity_size_mismatches() {
-        let http = test_http_client();
-        let history_hash = "a".repeat(64);
-        let context = context_with_history_ref(&history_hash);
-        let metadata = FinalSessionHistoryIdentity::new(
-            FinalSessionHistoryFramework::ClaudeCode,
-            hex::encode(Sha256::digest(b"sess-restore-plan")),
-            FinalSessionHistoryRefKind::Blob,
-            history_hash,
-            13,
-            "/home/user/.claude/projects/-home-user-workspace/session.jsonl",
-        )
-        .unwrap();
-        let restored_identity = RestoredSessionIdentity::from_final_metadata(
-            metadata,
-            "/home/user/.vm0/guest-agent/runs/previous/final-session-history-identity.json",
-            "/home/user/.vm0/guest-agent/runs/previous",
-        )
-        .expect("checkpointed final identity");
-        let reusable_sandbox = reusable_sandbox_with_identity(Some(restored_identity)).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::IdentityMismatch(Some(
-                        RestoredSessionIdentityMismatchReason::HistorySize
-                    )))
-                );
-            }
-            _ => panic!("reused identity with mismatched size should fall back to restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_falls_back_when_reused_identity_is_missing() {
-        let http = test_http_client();
-        let context = context_with_history_ref("history-hash-a");
-        let reusable_sandbox = reusable_sandbox_with_identity(None).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::MissingIdleIdentity)
-                );
-            }
-            _ => panic!("missing reused identity should fall back to restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_falls_back_when_finalizer_parked_without_identity() {
-        let http = test_http_client();
-        let context = context_with_history_ref("history-hash-a");
-        let reusable_sandbox = reusable_sandbox_parked_by_finalizer(None).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::MissingIdleIdentity)
-                );
-            }
-            _ => panic!("finalizer-parked missing identity should fall back to restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_classifies_history_hash_size_relationships() {
-        let http = test_http_client();
-        let requested_hash = "a".repeat(64);
-        let restored_hash = "b".repeat(64);
-        let cases = [
-            (
-                11,
-                RestoredSessionHistoryHashSizeRelationship::RequestedSmaller,
-            ),
-            (
-                12,
-                RestoredSessionHistoryHashSizeRelationship::RequestedEqual,
-            ),
-            (
-                13,
-                RestoredSessionHistoryHashSizeRelationship::RequestedLarger,
-            ),
-            (0, RestoredSessionHistoryHashSizeRelationship::SizeUnknown),
-            (
-                api_contracts::generated::constants::runners::RESUME_SESSION_HISTORY_MAX_BYTES + 1,
-                RestoredSessionHistoryHashSizeRelationship::SizeUnknown,
-            ),
-        ];
-
-        for (requested_size, expected_relationship) in cases {
-            let context = context_with_history_ref_and_size(&requested_hash, requested_size);
-            let restored_identity = final_metadata_identity(restored_hash.clone(), 12);
-            let reusable_sandbox = reusable_sandbox_with_identity(Some(restored_identity)).await;
-            let cancel = RunCancellationHandle::new();
-            let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-            let plan = build_session_history_restore_plan(
-                &http,
-                &SessionHistoryCpuPool::with_capacity(1),
-                &context,
-                &cancel,
-                SessionHistoryRestoreReuse {
-                    entry: Some(&reusable_sandbox),
-                    result: SandboxReuseResult::Reused,
-                },
-                &mut timing,
-                None,
-            );
-
-            match plan {
-                SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                    assert_eq!(
-                        fallback,
-                        Some(SessionHistoryRestoreFallback::IdentityMismatch(Some(
-                            RestoredSessionIdentityMismatchReason::HistoryHash(
-                                expected_relationship
-                            )
-                        )))
-                    );
-                }
-                _ => panic!("history hash mismatch should keep the prestarted restore plan"),
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_classifies_unverified_history_hash_size_as_unknown() {
-        let http = test_http_client();
-        let context = context_with_history_ref("history-hash-a");
-        let restored_identity = RestoredSessionIdentity::claude_code_for_test("history-hash-b");
-        let reusable_sandbox = reusable_sandbox_with_identity(Some(restored_identity)).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::IdentityMismatch(Some(
-                        RestoredSessionIdentityMismatchReason::HistoryHash(
-                            RestoredSessionHistoryHashSizeRelationship::SizeUnknown
-                        )
-                    )))
-                );
-            }
-            _ => panic!("unverified history hash mismatch should fall back to restore"),
-        }
-    }
-
-    #[test]
-    fn restore_plan_defers_hash_backed_history_for_non_reuse() {
-        let http = test_http_client();
-        let context = context_with_history_ref("history-hash-a");
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: None,
-                result: SandboxReuseResult::PoolMiss,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::DeferredHashBacked { fallback } => {
-                assert_eq!(fallback, Some(SessionHistoryRestoreFallback::NonReuse));
-            }
-            _ => panic!("non-reuse hash-backed history should defer materialization"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_classifies_session_identity_mismatch() {
-        let http = test_http_client();
-        let history_hash = "a".repeat(64);
-        let context = context_with_history_ref(&history_hash);
-        let restored_identity = RestoredSessionIdentity::new(
-            RestoredSessionFramework::ClaudeCode,
-            "sess-other",
-            crate::types::ResumeSessionHistoryRefKind::Blob,
-            history_hash,
-            Some(12),
-        );
-        let reusable_sandbox = reusable_sandbox_with_identity(Some(restored_identity)).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::IdentityMismatch(Some(
-                        RestoredSessionIdentityMismatchReason::SessionIdentity
-                    )))
-                );
-            }
-            _ => panic!("mismatched session identity should fall back to restore"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_plan_classifies_framework_mismatch() {
-        let http = test_http_client();
-        let history_hash = "a".repeat(64);
-        let context = context_with_history_ref(&history_hash);
-        let restored_identity = RestoredSessionIdentity::new(
-            RestoredSessionFramework::Codex,
-            "sess-restore-plan",
-            crate::types::ResumeSessionHistoryRefKind::Blob,
-            history_hash,
-            Some(12),
-        );
-        let reusable_sandbox = reusable_sandbox_with_identity(Some(restored_identity)).await;
-        let cancel = RunCancellationHandle::new();
-        let mut timing = RunnerPreSpawnTiming::start_after_claim();
-
-        let plan = build_session_history_restore_plan(
-            &http,
-            &SessionHistoryCpuPool::with_capacity(1),
-            &context,
-            &cancel,
-            SessionHistoryRestoreReuse {
-                entry: Some(&reusable_sandbox),
-                result: SandboxReuseResult::Reused,
-            },
-            &mut timing,
-            None,
-        );
-
-        match plan {
-            SessionHistoryRestorePlan::Prestarted { fallback, .. } => {
-                assert_eq!(
-                    fallback,
-                    Some(SessionHistoryRestoreFallback::IdentityMismatch(Some(
-                        RestoredSessionIdentityMismatchReason::Framework
-                    )))
-                );
-            }
-            _ => panic!("mismatched framework should fall back to restore"),
-        }
     }
 }
